more scripting changes
integrated cutscene with interactable objects, with a bit of trickery. all events now have a isCutscene() function which will return false unless it is a cutscene, and all have getSeq() and getDelay() which return empty arrays unless it is a cutscene. Finally, for moveEvents, it will continue doing the moveEvent until the movement is complete, regardless of the time given to do it, using a done() method in all events which returns true for all event subtypes other than moveEvent. First draft of the loadCutscene method written, untested.